1. strncat

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -15,7 +15,7 @@
 
 char *_strncat(char *dest, char *src, int n)
 {
-	size_t i, j;
+	int i, j;
 
 	i = 0;
 	j = 0;
